Lift userinput usestate and handle change functions to App.js component
