+++ src/components/UserInput.jsx
@@ -1,21 +1,4 @@
-import React, { useState } from "react";
-
-const UserInput = () => {
-  const [userInput, setUserInput] = useState({
-    InitialInvestment: 1000,
-    annualInvestment: 1200,
-    expectedReturn: 6,
-    duration: 5,
-  });
-
-  const handleChange = (inputIdentifier, newValue) => {
-    setUserInput((prevUserInput) => {
-      return {
-        ...prevUserInput,
-        [inputIdentifier]: newValue,
-      };
-    });
-  };
+const UserInput = ({onChange, userInput}) => { 
 
   return (
     <section id="user-input">
@@ -26,7 +9,7 @@ const UserInput = () => {
             type="number"
             required
             value={userInput.InitialInvestment}
-            onChange={(e) => handleChange("InitialInvestment", e.target.value)}
+            onChange={(e) => onChange("InitialInvestment", e.target.value)}
           />
         </p>
         <p>
@@ -35,7 +18,7 @@ const UserInput = () => {
             type="number"
             required
             value={userInput.annualInvestment}
-            onChange={(e) => handleChange("annualInvestment", e.target.value)}
+            onChange={(e) => onChange("annualInvestment", e.target.value)}
           />
         </p>
       </div>
@@ -47,13 +30,13 @@ const UserInput = () => {
             type="number"
             required
             value={userInput.expectedReturn}
-            onChange={(e) => handleChange("expectedReturn", e.target.value)}
+            onChange={(e) => onChange("expectedReturn", e.target.value)}
           />
         </p>
         <p>
           <label>Duration</label>
           <input type="number" required value={userInput.duration}
-            onChange={(e) => handleChange("duration", e.target.value)}/>
+            onChange={(e) => onChange("duration", e.target.value)}/>
         </p>
       </div>
     </section>
